Isolated context menu creation logic for testability Added:   - ContextMenuFactory: handles ui creation and binding of context     menus, handles single/multi-file ops.
